- renamed workspace to schedule and recreated workspace - GUI: fixed bug with not correct changing workspace in - GUI: all functions of tab 'Workspaces' mowed to tag 'Edit - views: context filling moved to separate method

+++ workspace_editor/views.py
@@ -1,5 +1,5 @@
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render
+import logging
 
 from workspace_editor.forms import ScheduleForm, EventCreateForm, EventEditForm, CompilationCreateForm
 from workspace_editor.models import Workspace
@@ -7,12 +7,12 @@ from workspace_editor.serializers import WorkspaceSerializer
 from django.views.generic import View, TemplateView, CreateView, FormView, DetailView, ListView
 from django.contrib.auth import authenticate, login, logout
 from django.shortcuts import render, redirect
+from django.http import HttpResponseRedirect, HttpResponse
+
 from django.urls import reverse_lazy, reverse
 from django.core.paginator import Paginator
 from django.core.mail import send_mail
 from django.http import JsonResponse
-from django.conf import settings
-from django.db.models import Q
 from .models import *
 import calendar
 from .forms import *
@@ -20,35 +20,42 @@ import datetime
 
 from loader.models import *
 
-import requests
+log = logging.getLogger(__name__)
+
 
 
 @login_required
-def homeView(request):
-    context = {}
-    create_event_form = EventCreateForm()
-    edit_event_form = EventEditForm()
+def home(request):
 
-    selected_workspace_id = None
     workspace = None
     schedule = None
+    create_event_form = None
+    edit_event_form = None
+    first_workspace = False
+    workspace_id = int(request.GET['workspace_id'])
+
 
     # Events stuff
-    if "selected_workspace_id" in request.GET:
-        selected_workspace_id = request.GET["selected_workspace_id"]
-        first_workspace = True
+    if workspace_id and Workspace.objects.filter(workspace_id=workspace_id) \
+            .filter(Q(owner=request.user.pk) | Q(visible_for=request.user)).exists():
+        workspace = Workspace.objects.get(workspace_id=workspace_id)
+    elif workspace_id is None:
+        log.info(f"You don't have available workspaces")
     else:
+        log.warning(f"Workspace with id workspace_id={workspace_id} doesn't exists or you don't have access to it")
         first_workspace = Workspace.objects.filter(Q(owner=request.user.pk) | Q(visible_for=request.user)).first()
         if first_workspace:
-            selected_workspace_id = first_workspace.workspace_id
+            workspace_id = first_workspace.workspace_id
+            workspace = Workspace.objects.get(workspace_id=workspace_id)
 
-    if selected_workspace_id:
-        workspace = Workspace.objects.get(workspace_id=selected_workspace_id)
+    if workspace:
         schedule = Schedule.objects.get(schedule_id=workspace.schedule_id)
+
 
     if request.POST:
         if request.POST['action'] == 'create':
             form = WorkspaceForm(request.POST)
+
             if form.is_valid():
                 form.set_owner(request.user)
                 workspace = form.save()
@@ -65,7 +72,7 @@ def homeView(request):
             else:
                 print(form.errors.as_data())
 
-        if request.POST['action'] == 'edit':
+        elif workspace and request.POST['action'] == 'edit':
             form = WorkspaceSettingsForm(request.POST)
             if form.is_valid():
                 form.save(commit=True)
@@ -73,16 +80,16 @@ def homeView(request):
                 print(form.errors.as_data())
 
 
-        if request.POST['action'] == 'delete':
+        elif workspace and request.POST['action'] == 'delete':
             workspace = Workspace.objects.get(workspace_id=request.POST["workspace_id"])
             if workspace.owner == request.user:
                 workspace.delete()
                 first_workspace = Workspace.objects.filter(Q(owner=request.user.pk) | Q(visible_for=request.user)).first()
                 if first_workspace:
-                    selected_workspace_id = first_workspace.workspace_id
+                    workspace_id = first_workspace.workspace_id
 
 
-        if request.POST['action'] == "create_event":
+        elif workspace and request.POST['action'] == "create_event":
             form = EventCreateForm(request.POST)
             if form.is_valid():
                 form.safe_copied_post(workspace.scheduled_compilation_id)
@@ -94,7 +101,7 @@ def homeView(request):
                 create_event_form = form
 
 
-        if request.POST['action'] == "edit_event":
+        elif workspace and request.POST['action'] == "edit_event":
             form = EventEditForm(request.POST)
             if form.is_valid():
                 form.save()
@@ -104,7 +111,7 @@ def homeView(request):
                 edit_event_form = form
 
 
-        if request.POST['action'] == 'create_compilation':
+        elif workspace and request.POST['action'] == 'create_compilation':
             form = CompilationCreateForm(request.POST)
             if form.is_valid():
                  form.save()
@@ -127,58 +134,68 @@ def homeView(request):
         #             selected_compilation_id = firstCompilation.id
 
 
+    context = prepare_context(request, schedule, create_event_form, edit_event_form, workspace_id)
+
+    # return HttpResponseRedirect(reverse('workspace', args=(id,)), context)
+    # return HttpResponseRedirect("workspace.html")
+
+    return render(request, "workspace.html", context)
 
 
+def prepare_context(request, schedule, create_event_form, edit_event_form, workspace_id):
+    context = {}
 
-    # workspace stuff
     queryset_visible = Workspace.objects.filter(Q(owner=request.user.pk) | Q(visible_for=request.user))
+    context["my_workspaces"] = queryset_visible
+
     queryset_editable = Workspace.objects.filter(Q(owner=request.user.pk) | Q(editable_by=request.user))
     context["workspaces"] = WorkspaceSerializer(queryset_editable, many=True).data
 
     context["createform"] = WorkspaceForm()
-    context["settingsform"] = WorkspaceSettingsForm(initial={"user_id": request.user.pk, "owner": request.user})
-    context["my_workspaces"] = queryset_visible
 
-    # TODO: make table compilationsOwners with compilation_id, owner and visible_for for looking for
-    #  only current user compilations
-    context["my_compilations"] = Compilation.objects.all()
+    context["selected_workspace_id"] = workspace_id
 
-    if first_workspace and schedule:
+    if schedule:
+        context["settingsform"] = WorkspaceSettingsForm(initial={"user_id": request.user.pk, "owner": request.user})
+
+        # TODO: make table compilationsOwners with compilation_id, owner and visible_for for looking for
+        #  only current user compilations
+        context["my_compilations"] = Compilation.objects.all()
         context["schedule_events"] = get_events_for_schedule(schedule.schedule_id)
         context["selected_schedule_id"] = int(schedule.schedule_id)
-    context["event_createform"] = create_event_form
-    context["event_editform"] = edit_event_form
+        if create_event_form:
+            context["event_createform"] = create_event_form
+        if create_event_form:
+            context["edit_event_form"] = edit_event_form
 
+        all_post_entries = PostEntry.objects.all()
+        # TODO: change number of items and make slider
+        paginator = Paginator(all_post_entries, 8)
 
-    all_post_entries = PostEntry.objects.all()
-    # TODO: change number of items and make slider
-    paginator = Paginator(all_post_entries, 8)
+        page_number = request.GET.get('page')
+        post_list = paginator.get_page(page_number)
+        context['post_list'] = post_list
 
-    page_number = request.GET.get('page')
-    post_list = paginator.get_page(page_number)
-    context['post_list'] = post_list
-
-    return render(request, "home.html", context)
-
+    return context
 
 
 def get_events_for_schedule(selected_schedule_id):
     dates_to_events = dict()
     count = 0
     date = datetime.date.today()
+    days_to_show = 7
 
     while (True):
         day_of_week = calendar.day_name[date.weekday()]
 
         if Event.objects.filter(schedule_id=selected_schedule_id, start_date__year=date.year,
                                 start_date__month=date.month, start_date__day=date.day).exists():
-            events = Event.objects.filter(schedule_id=selected_schedule_id,start_date__year=date.year,
+            events = Event.objects.filter(schedule_id=selected_schedule_id, start_date__year=date.year,
                                           start_date__month=date.month, start_date__day=date.day) \
                 .order_by('start_date')
             events_to_posts = dict()
             for event in events:
                 events_to_posts[event] = PostEntry.objects.get(id=event.post_id)
-
 
             dates_to_events[(date, day_of_week)] = events_to_posts
         else:
@@ -186,7 +203,8 @@ def get_events_for_schedule(selected_schedule_id):
 
         date += datetime.timedelta(days=1)
         count += 1
-        if count >= 7 and not Event.objects.filter(start_date__range=[date, date + datetime.timedelta(days=365)]).exists():
+        if count >= days_to_show \
+                and not Event.objects.filter(start_date__range=[date, date + datetime.timedelta(days=365)]).exists():
             break
 
     return dates_to_events
